refatoring(project): portfolio write,edit module
