+++ app/(admin)/admin/write/WritePortfolioForm.module.css
@@ -4,16 +4,6 @@
   @apply mx-auto w-full max-w-[1100px] px-4 pb-14 pt-8 sm:px-6;
 }
 
-.panel {
-  @apply overflow-hidden rounded-[18px];
-  background: rgb(255, 248, 240);
-  border: 1px solid rgba(255, 255, 255, 0.95);
-  box-shadow:
-    0 20px 52px rgba(94, 72, 34, 0.12),
-    0 2px 10px rgba(94, 72, 34, 0.08),
-    inset 0 1px 0 rgba(255, 255, 255, 0.92);
-}
-
 .header {
   @apply flex flex-col gap-3 border-b px-5 py-5 sm:px-8;
   border-color: rgba(130, 102, 64, 0.16);
@@ -26,17 +16,6 @@
 
 .headerTop {
   @apply flex items-center justify-between gap-3;
-}
-
-.backLink {
-  @apply inline-flex rounded-full px-3 py-1 text-sm font-semibold transition-colors;
-  border: 1px solid rgba(144, 114, 74, 0.34);
-  background: rgba(250, 241, 227, 0.78);
-  color: rgba(80, 58, 33, 0.9);
-}
-
-.backLink:hover {
-  background: rgba(255, 248, 236, 0.95);
 }
 
 .statusChip {
@@ -54,102 +33,4 @@
 .subtitle {
   @apply m-0 text-sm;
   color: rgba(70, 55, 40, 0.74);
-}
-
-.form {
-  @apply flex flex-col gap-5 px-5 py-6 sm:px-8 sm:py-8;
-}
-
-.grid {
-  @apply grid grid-cols-1 gap-4 sm:grid-cols-2;
-}
-
-.fieldGroup {
-  @apply flex flex-col gap-1.5;
-}
-
-.label {
-  @apply text-sm font-semibold;
-  color: rgba(46, 33, 19, 0.88);
-}
-
-.input {
-  @apply w-full rounded-md border-0 px-3 py-2 text-sm focus:outline-none focus:ring-1 focus:ring-offset-0;
-  background: rgba(255, 255, 255, 0.82);
-  color: rgba(46, 35, 24, 0.9);
-}
-
-.input[type="file"] {
-  @apply py-2;
-}
-
-.textarea {
-  @apply min-h-24 w-full rounded-md border-0 px-3 py-2 text-sm leading-relaxed focus:outline-none focus:ring-1 focus:ring-offset-0;
-  background: rgba(255, 255, 255, 0.82);
-  color: rgba(46, 35, 24, 0.9);
-}
-
-.contentsArea {
-  @apply min-h-[360px] font-mono text-[0.88rem] leading-6;
-}
-
-.actions {
-  @apply mt-1 flex flex-wrap items-center gap-2;
-}
-
-.primaryButton {
-  @apply inline-flex w-fit items-center justify-center rounded-full px-4 py-2 text-sm font-semibold transition-colors;
-  border: 1px solid rgba(144, 114, 74, 0.36);
-  background: rgba(250, 241, 227, 0.92);
-  color: rgba(78, 56, 31, 0.96);
-}
-
-.primaryButton:hover {
-  background: rgba(255, 248, 236, 1);
-}
-
-.primaryButton:disabled {
-  opacity: 0.6;
-  cursor: not-allowed;
-}
-
-.secondaryButton {
-  @apply inline-flex w-fit items-center justify-center rounded-full px-4 py-2 text-sm font-semibold transition-colors;
-  border: 1px solid rgba(144, 114, 74, 0.26);
-  background: rgba(255, 255, 255, 0.72);
-  color: rgba(78, 56, 31, 0.88);
-}
-
-.secondaryButton:hover {
-  background: rgba(255, 255, 255, 0.92);
-}
-
-.cancelLink {
-  @apply inline-flex w-fit items-center justify-center rounded-full px-4 py-2 text-sm font-semibold transition-colors;
-  border: 1px solid rgba(144, 114, 74, 0.26);
-  background: rgba(255, 255, 255, 0.56);
-  color: rgba(78, 56, 31, 0.84);
-}
-
-.cancelLink:hover {
-  background: rgba(255, 255, 255, 0.82);
-}
-
-.alert {
-  @apply m-0 rounded-md px-3 py-2 text-sm font-medium;
-  color: rgba(130, 36, 36, 0.92);
-  background: rgba(255, 231, 231, 0.86);
-  border: 1px solid rgba(199, 110, 110, 0.4);
-}
-
-.alertSuccess {
-  color: rgba(28, 95, 64, 0.95);
-  background: rgba(223, 247, 236, 0.9);
-  border-color: rgba(99, 170, 138, 0.42);
-}
-
-@media (max-width: 640px) {
-  .contentsArea {
-    min-height: 280px;
-  }
 }
