+++ app/components/portfolio/PortfolioForm.module.css
@@ -1,0 +1,158 @@
+@reference "../../globals.css";
+
+.form {
+  @apply flex flex-col gap-6 px-5 py-6 sm:px-8 sm:py-8;
+}
+
+/* 메타데이터 섹션 */
+.metaSection {
+  @apply border-b pb-6;
+  border-color: rgba(130, 102, 64, 0.12);
+}
+
+/* 콘텐츠 섹션 */
+.contentSection {
+  @apply pt-2;
+}
+
+.grid {
+  @apply grid grid-cols-1 gap-4 sm:grid-cols-2 lg:grid-cols-3;
+}
+
+.fieldGroup {
+  @apply flex flex-col gap-1.5;
+}
+
+.label {
+  @apply text-sm font-semibold;
+  color: rgba(46, 33, 19, 0.88);
+}
+
+.input {
+  @apply w-full rounded-md border-0 px-3 py-2 text-sm focus:outline-none focus:ring-1 focus:ring-offset-0;
+  background: rgba(255, 255, 255, 0.82);
+  color: rgba(46, 35, 24, 0.9);
+  border-color: transparent;
+}
+
+.input:focus {
+  border-color: rgba(144, 114, 74, 0.3);
+  box-shadow: 0 0 0 3px rgba(255, 200, 120, 0.15);
+}
+
+.select {
+  @apply w-full rounded-md border-0 px-3 py-2 text-sm focus:outline-none focus:ring-1 focus:ring-offset-0;
+  background: rgba(255, 255, 255, 0.82);
+  color: rgba(46, 35, 24, 0.9);
+  cursor: pointer;
+  appearance: none;
+  background-image: url("data:image/svg+xml;charset=UTF-8,%3csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 24 24' fill='none' stroke='currentColor' stroke-width='2' stroke-linecap='round' stroke-linejoin='round'%3e%3cpolyline points='6 9 12 15 18 9'%3e%3c/polyline%3e%3c/svg%3e");
+  background-repeat: no-repeat;
+  background-position: right 8px center;
+  background-size: 20px;
+  padding-right: 32px;
+}
+
+.select:focus {
+  border-color: rgba(144, 114, 74, 0.3);
+  box-shadow: 0 0 0 3px rgba(255, 200, 120, 0.15);
+}
+
+.select option {
+  background: rgba(255, 248, 240, 1);
+  color: rgba(46, 35, 24, 0.9);
+}
+
+.textarea {
+  @apply min-h-24 w-full rounded-md border-0 px-3 py-2 text-sm leading-relaxed focus:outline-none focus:ring-1 focus:ring-offset-0;
+  background: rgba(255, 255, 255, 0.82);
+  color: rgba(46, 35, 24, 0.9);
+}
+
+.contentsArea {
+  @apply min-h-[360px] font-mono text-[0.88rem] leading-6;
+}
+
+.actions {
+  @apply mt-8 flex flex-wrap items-center gap-2 border-t pt-6;
+  border-color: rgba(130, 102, 64, 0.12);
+}
+
+.primaryButton {
+  @apply inline-flex w-fit items-center justify-center rounded-full px-4 py-2 text-sm font-semibold transition-colors;
+  border: 1px solid rgba(144, 114, 74, 0.36);
+  background: rgba(250, 241, 227, 0.92);
+  color: rgba(78, 56, 31, 0.96);
+}
+
+.primaryButton:hover {
+  background: rgba(255, 248, 236, 1);
+}
+
+.primaryButton:disabled {
+  opacity: 0.6;
+  cursor: not-allowed;
+}
+
+.secondaryButton {
+  @apply inline-flex w-fit items-center justify-center rounded-full px-4 py-2 text-sm font-semibold transition-colors;
+  border: 1px solid rgba(144, 114, 74, 0.26);
+  background: rgba(255, 255, 255, 0.72);
+  color: rgba(78, 56, 31, 0.88);
+}
+
+.secondaryButton:hover {
+  background: rgba(255, 255, 255, 0.92);
+}
+
+.cancelLink {
+  @apply inline-flex w-fit items-center justify-center rounded-full px-4 py-2 text-sm font-semibold transition-colors;
+  border: 1px solid rgba(144, 114, 74, 0.26);
+  background: rgba(255, 255, 255, 0.56);
+  color: rgba(78, 56, 31, 0.84);
+}
+
+.cancelLink:hover {
+  background: rgba(255, 255, 255, 0.82);
+}
+
+.messageWrap {
+  @apply mt-2;
+}
+
+.successMessage,
+.errorMessage {
+  @apply flex items-start gap-2 rounded-lg border px-3 py-2 text-sm;
+}
+
+.messageIcon {
+  @apply mt-0.5 inline-flex h-5 w-5 shrink-0 items-center justify-center rounded-full text-xs font-bold;
+}
+
+.successMessage {
+  background: rgba(230, 249, 238, 0.75);
+  border-color: rgba(44, 128, 84, 0.32);
+  color: rgba(29, 98, 63, 0.95);
+}
+
+.successMessage .messageIcon {
+  background: rgba(44, 128, 84, 0.16);
+  color: rgba(29, 98, 63, 0.95);
+}
+
+.errorMessage {
+  background: rgba(255, 241, 236, 0.8);
+  border-color: rgba(178, 85, 63, 0.34);
+  color: rgba(132, 50, 31, 0.95);
+}
+
+.errorMessage .messageIcon {
+  background: rgba(178, 85, 63, 0.16);
+  color: rgba(132, 50, 31, 0.95);
+}
+
+@media (max-width: 640px) {
+  .contentsArea {
+    min-height: 280px;
+  }
+}
